Term length changed to 2 years, optionally 1 year
Changes to the charter motivated by potential candidates for the board. Three years, as introduced in 2023, feels too long for people with temporary contracts, as most of UiO carpentry members have. The proposed changes try to accommodate this and allow for more flexibility, so that members with less than two years left at UiO are nevertheless eligible to join the board.